refactor : 모임 아이템 컴포넌트 디렉토리 위치 변경으로 코드 수정
